examples/fastapi_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
asyncpg==0.30.0
redis==5.2.1
loguru==0.7.3
prometheus-client==0.21.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0

